ems_rest_api/requirements.txt
asgiref==3.6.0
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
cryptography==41.0.1
dataclasses==0.6
Django==4.1.9
django-cors-headers==4.0.0
django-utils-six==2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djongo==1.3.6
dnspython==2.3.0
idna==3.4
msgpack==1.0.5
ndg-httpsclient==0.5.1
pusher==3.3.2
pyasn1==0.5.0
pycparser==2.21
PyJWT==2.7.0
pymongo==3.12.3
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pytz==2023.3
redis==4.5.5
requests==2.31.0
six==1.16.0
sqlparse==0.2.4
tzdata==2023.3
urllib3==2.0.3
